Refactor Issue into models model, and DataclassSerializer into utils
