+++ jira_cli/main.py
@@ -1,18 +1,10 @@
 import collections.abc
-import copy
-import dataclasses
-from dataclasses import dataclass, field
 import datetime
-import decimal
-import enum
 import json
 import logging
 import os
-import textwrap
 import urllib3
 
-import arrow
-import dictdiffer
 import pandas as pd
 
 import jira as mod_jira
@@ -20,226 +12,16 @@ import jira as mod_jira
 from tabulate import tabulate
 from tqdm import tqdm
 
+from jira_cli.models import Issue
+
 
 CUSTOM_FIELD_EPIC_LINK = 'customfield_14182'
 CUSTOM_FIELD_EPIC_NAME = 'customfield_14183'
 CUSTOM_FIELD_ESTIMATE = 'customfield_10002'
 
 
-class IssueStatus(enum.Enum):
-    Backlog = 'Backlog'
-    ToDo = 'To Do'
-    InProgress = 'In Progress'
-    StoryInProgress = 'Story in Progress'
-    EpicInProgress = 'Epic in Progress'
-    EpicWithSquad = 'Epic with Squad'
-    EpicReadyforSquad = 'Epic Ready for Squad'
-    InRelease = 'In Release'
-    Done = 'Done'
-    StoryDone = 'Story Done'
-    EpicDone = 'Epic Done'
-    Closed = 'Closed'
-    RiskClosed = 'Risk Closed'
-    RiskIdentified = 'Risk Identified'
-    NotAssessed = 'Not Assessed'
-    Resolved = 'Resolved'
-    Accepted = 'Accepted'
-
-
 logger = logging.getLogger('jira')
 
-
-@dataclass
-class DataclassSerializer:
-    @classmethod
-    def deserialize(cls, attrs: dict) -> object:
-        """
-        Deserialize JIRA API dict to dataclass
-        Support decimal, date/datetime, enum & set
-        """
-        data = copy.deepcopy(attrs)
-
-        for f in dataclasses.fields(cls):
-            v = attrs.get(f.name)
-            if v is None:
-                continue
-
-            if dataclasses.is_dataclass(f.type):
-                data[f.name] = f.type.deserialize(v)
-            elif f.type is decimal.Decimal:
-                data[f.name] = decimal.Decimal(v)
-            elif issubclass(f.type, enum.Enum):
-                # convert string to Enum instance
-                data[f.name] = f.type(v)
-            elif f.type is datetime.date:
-                data[f.name] = arrow.get(v).datetime.date()
-            elif f.type is datetime.datetime:
-                data[f.name] = arrow.get(v).datetime
-            elif f.type is set:
-                data[f.name] = set(v)
-
-        return cls(**data)
-
-    def serialize(self) -> dict:
-        """
-        Serialize dataclass to JIRA API dict
-        Support decimal, date/datetime, enum & set
-        Include only fields with repr=True (dataclass.field default)
-        """
-        data = {}
-
-        for f in dataclasses.fields(self):
-            if f.repr is False:
-                continue
-
-            v = self.__dict__.get(f.name)
-
-            if v is None:
-                data[f.name] = None
-            elif dataclasses.is_dataclass(f.type):
-                data[f.name] = v.serialize()
-            elif isinstance(v, decimal.Decimal):
-                data[f.name] = str(v)
-            elif issubclass(f.type, enum.Enum):
-                # convert Enum to raw string
-                data[f.name] = v.value
-            elif isinstance(v, (datetime.date, datetime.datetime)):
-                data[f.name] = v.isoformat()
-            elif isinstance(v, set):
-                data[f.name] = sorted(list(v))
-            else:
-                data[f.name] = v
-
-        return data
-
-
-# pylint: disable=too-many-instance-attributes
-@dataclass
-class Issue(DataclassSerializer):
-    assignee: str
-    created: datetime.datetime
-    creator: str
-    description: str
-    fixVersions: set
-    issuetype: str
-    key: str
-    labels: set
-    lastViewed: datetime.datetime
-    priority: str
-    project: str
-    reporter: str
-    status: IssueStatus
-    summary: str
-    updated: datetime.datetime
-    estimate: int = field(default=None)
-    epic_ref: str = field(default=None)
-    epic_name: str = field(default=None)
-
-    # local-only dict which represents serialized Issue last seen on JIRA server
-    # this property is not written to cache and is created at runtme from diff_to_original
-    original: dict = field(default=None, repr=False)
-
-    # patch of current Issue to dict last seen on JIRA server
-    diff_to_original: list = field(default=None, repr=False)
-
-    @property
-    def is_open(self) -> bool:
-        if self.is_inprogress or self.is_todo:
-            return True
-        elif self.is_done or self.is_closed:
-            return False
-        else:
-            raise AttributeError('Issue cannot be determined open or closed!')
-
-    @property
-    def is_todo(self) -> bool:
-        if self.status in (IssueStatus.Backlog, IssueStatus.ToDo, IssueStatus.RiskIdentified,
-                           IssueStatus.NotAssessed, IssueStatus.EpicReadyforSquad):
-            return True
-        return False
-
-    @property
-    def is_inprogress(self) -> bool:
-        if self.status in (IssueStatus.InProgress, IssueStatus.InRelease, IssueStatus.Accepted,
-                           IssueStatus.EpicInProgress, IssueStatus.StoryInProgress,
-                           IssueStatus.EpicWithSquad):
-            return True
-        return False
-
-    @property
-    def is_done(self) -> bool:
-        if self.status in (IssueStatus.Done, IssueStatus.StoryDone, IssueStatus.EpicDone,
-                           IssueStatus.Resolved):
-            return True
-        return False
-
-    @property
-    def is_closed(self) -> bool:
-        if self.is_done or self.status in (IssueStatus.Closed, IssueStatus.RiskClosed):
-            return True
-        return False
-
-    @classmethod
-    def deserialize(cls, attrs: dict) -> object:
-        '''
-        Deserialize a dict into an Issue object. Inflate the original Jira issue from the
-        diff_to_original property
-        '''
-        # deserialize supplied dict into an Issue object
-        issue = super().deserialize(attrs)
-
-        # pylint: disable=no-member
-        if issue.diff_to_original is None:
-            issue.diff_to_original = []
-
-        # apply the diff_to_original patch to the serialized version of the issue, which recreates
-        # the issue dict as last seen on the JIRA server
-        # pylint: disable=no-member
-        issue.original = dictdiffer.patch(issue.diff_to_original, issue.serialize())
-
-        return issue
-
-    def serialize(self) -> dict:
-        '''
-        Serialize this Issue object to a dict. Generate the diff_to_original field which enables
-        local changes to be written to the offline cache file.
-        '''
-        # serialize self (Issue object) into a dict
-        data = super().serialize()
-
-        if self.original:
-            # if this Issue object has the original property set, render the diff between self and
-            # the original property. This is written to storage to track changes made offline.
-            data['diff_to_original'] = list(
-                dictdiffer.diff(data, self.original, ignore=set(['diff_to_original']))
-            )
-
-        return data
-
-    def __str__(self):
-        '''Pretty print this Issue'''
-        if self.issuetype == 'Epic':
-            epicdetails = ('Epic Short Name', f'{self.epic_name}')
-        else:
-            epicdetails = ('Epic Ref', f'{self.epic_ref}')
-
-        return tabulate([
-            ('Summary', f'[{self.key}] {self.summary}'),
-            ('Type', self.issuetype),
-            epicdetails,
-            ('Status', self.status.value),
-            ('Priority', self.priority),
-            ('Assignee', self.assignee),
-            ('Estimate', self.estimate),
-            ('Description', '\n'.join(textwrap.wrap(self.description, width=100))),
-            ('Fix Version', tabulate([('-', v) for v in self.fixVersions], tablefmt='plain')),
-            ('Labels', tabulate([('-', l) for l in self.labels], tablefmt='plain')),
-            ('Reporter', self.reporter),
-            ('Creator', self.creator),
-            ('Created', self.created),
-            ('Updated', self.updated),
-            ('LastViewed', self.lastViewed),
-        ])
 
 
 class Jira(collections.abc.MutableMapping):
